additional checks in backend + toast fix + avoid username conflicts

+++ backend/src/relationships/relationships.controller.ts
@@ -1,6 +1,16 @@
-import { Controller, Get, Post, Body, Delete, Param, UseGuards } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Delete,
+  Param,
+  UseGuards,
+  BadRequestException,
+} from '@nestjs/common';
 import { RelationshipsService } from './relationships.service';
 import { Relationship } from './interfaces/relationship.interface';
+import { UsersService } from '../users/users.service';
 import { CreateRelationshipDto } from './dto/createRelationship.dto';
 import { ValidateRelationshipDto } from './dto/validateRelationship.dto';
 import { RemoveRelationshipDto } from './dto/removeRelationship.dto';
@@ -10,10 +20,17 @@ import { JwtTwoFactorGuard } from 'src/auth/guards/jwtTwoFactor.guard';
 @ApiTags('Relationship')
 @Controller('relationships')
 export class RelationshipsController {
-  constructor(private readonly relationshipService: RelationshipsService) {}
+  constructor(
+    private readonly relationshipService: RelationshipsService,
+    private readonly userService: UsersService,
+  ) {}
+
+  //GETTERS
 
   @Get('/friendships/:id')
   async getUserFriendships(@Param('id') id: number): Promise<Relationship[]> {
+    const currentUser = await this.userService.getUserbyId(id);
+    if (!currentUser) throw new BadRequestException("user doesn't exist");
     return await this.relationshipService.getUserFriendships(id);
   }
 
@@ -21,24 +38,40 @@ export class RelationshipsController {
   async getAllUserFriendships(
     @Param('id') id: number,
   ): Promise<Relationship[]> {
+    const currentUser = await this.userService.getUserbyId(id);
+    if (!currentUser) throw new BadRequestException("user doesn't exist");
     return await this.relationshipService.getAllUserFriendships(id);
   }
 
   @Get('/blocked/:id')
   async getUserBlocked(@Param('id') id: number): Promise<Relationship[]> {
+    const currentUser = await this.userService.getUserbyId(id);
+    if (!currentUser) throw new BadRequestException("user doesn't exist");
     return await this.relationshipService.getUserBlocked(id);
   }
 
   @Get('/blocked-by/:id')
   async getBlockedByUser(@Param('id') id: number): Promise<Relationship[]> {
+    const currentUser = await this.userService.getUserbyId(id);
+    if (!currentUser) throw new BadRequestException("user doesn't exist");
     return await this.relationshipService.getBlockedByUser(id);
   }
+
+  // SAVE-UPDATE-DELETE
 
   @Post()
   @UseGuards(JwtTwoFactorGuard)
   async saveRelationship(
     @Body() newRelationship: CreateRelationshipDto,
   ): Promise<Relationship> {
+    const user1 = await this.userService.getUserbyId(
+      newRelationship.requesterId,
+    );
+    if (!user1) throw new BadRequestException("user doesn't exist");
+    const user2 = await this.userService.getUserbyId(
+      newRelationship.adresseeId,
+    );
+    if (!user2) throw new BadRequestException("user doesn't exist");
     return await this.relationshipService.saveRelationship(newRelationship);
   }
 
@@ -47,9 +80,11 @@ export class RelationshipsController {
   async validateRelationship(
     @Body() toMajRelationship: ValidateRelationshipDto,
   ) {
-    return await this.relationshipService.validateRelationship(
-      toMajRelationship,
+    const relationship = await this.relationshipService.getRelationship(
+      toMajRelationship.requesterId,
+      toMajRelationship.adresseeId,
     );
+    return await this.relationshipService.validateRelationship(relationship);
   }
 
   @Delete()
@@ -57,6 +92,12 @@ export class RelationshipsController {
   async removeRelationship(
     @Body() toRemoveRelationship: RemoveRelationshipDto,
   ) {
+    const relationship = await this.relationshipService.getRelationship(
+      toRemoveRelationship.userId1,
+      toRemoveRelationship.userId2,
+    );
+    if (!relationship)
+      throw new BadRequestException("relationship doesn't exist");
     return await this.relationshipService.removeRelationship(
       toRemoveRelationship,
     );
